Move spec helper methods to their own file
